pyproject.toml
[tool.poetry]
name = "xrpl-sidechain-cli"
version = "1.0.0-beta.0"
description = "A CLI that helps you set up an XRPL sidechain."
readme = "README.md"
repository = "https://github.com/xpring-eng/sidechain-cli"
authors = ["Mayukha Vadari <mvadari@ripple.com>"]
keywords = [
  "xrp",
  "xrpl",
  "cryptocurrency",
]
license = "MIT"
include = [
  "LICENSE",
]
packages = [
  { include = "sidechain_cli" },
]

[tool.poetry.scripts]
sidechain-cli = 'sidechain_cli.main:main'

[tool.poetry.dependencies]
python = "^3.7.1"
click = "^8.1.3"
tabulate = "^0.8.9"
xrpl-py = "^1.7.0b0"
httpx = "^0.18.1"
websockets = "^10.3"
Jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3"
flake8-black = "^0.3.3"
flake8-docstrings = "^1.5.0"
mypy = "^0"
isort = "^5.7.0"
flake8-isort = "^4.0.0"
flake8-annotations = "^2.5.0"
flake8-absolute-import = "^1.0"
darglint = "^1.5.8"
types-tabulate = "^0.8.9"
coverage = "^6.4.4"
pytest = "^7.1.3"

[tool.isort]
# Make sure that isort's settings line up with black
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
